# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Vega API
  description: REST API for Vega, an AI agent orchestration framework.
  version: "1.0.0"
servers:
  - url: https://your-vega.example.com
    description: Production
paths:
...
  /api/v1/mcp/servers:
    get:
      tags: [MCP]
      summary: List connected MCP servers
      operationId: listMCPServers
      responses:
        "200":
          description: Array of MCP server statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MCPServerResponse"
    post:
      tags: [MCP]
      summary: Connect an MCP server
      operationId: connectMCPServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectMCPRequest"
      responses:
        "200":
          description: Server connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectMCPResponse"
        "409":
          description: Server already connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Connection failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectMCPResponse"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string
...
    MCPServerResponse:
      type: object
      properties:
        name:
          type: string
        connected:
          type: boolean
        disabled:
          type: boolean
        transport:
          type: string
        url:
          type: string
        command:
          type: string
        tools:
          type: array
          items:
            type: string
...
    ConnectMCPRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        env:
          type: object
          additionalProperties:
            type: string
        transport:
          type: string
          enum: [stdio, http, sse]
        command:
          type: string
        args:
          type: array
          items:
            type: string
        url:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
        timeout:
          type: integer
          description: Timeout in seconds
    ConnectMCPResponse:
      type: object
      properties:
        name:
          type: string
        connected:
          type: boolean
        tools:
          type: array
          items:
            type: string
        error:
          type: string
...